+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/client/DeliveryClient.java
@@ -6,6 +6,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import ru.yandex.practicum.dto.delivery.DeliveryDto;
 import ru.yandex.practicum.dto.order.OrderDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @FeignClient(name = "delivery")
@@ -15,7 +16,7 @@ public interface DeliveryClient {
     DeliveryDto planDelivery(@RequestBody DeliveryDto delivery);
 
     @PostMapping("/api/v1/delivery/cost")
-    Double deliveryCost(@RequestBody OrderDto order);
+    BigDecimal deliveryCost(@RequestBody OrderDto order);
 
     @PostMapping("/api/v1/delivery/successful")
     void deliverySuccessful(@RequestBody UUID orderId);
