+++ commerce/order/src/main/java/ru/yandex/practicum/order/client/DeliveryOrderFeignClient.java
@@ -6,6 +6,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import ru.yandex.practicum.dto.delivery.DeliveryDto;
 import ru.yandex.practicum.dto.order.OrderDto;
 
+import java.math.BigDecimal;
+
 @FeignClient(name = "delivery")
 public interface DeliveryOrderFeignClient {
 
@@ -13,5 +15,5 @@ public interface DeliveryOrderFeignClient {
     DeliveryDto planDelivery(@RequestBody DeliveryDto delivery);
 
     @PostMapping("/api/v1/delivery/cost")
-    Double deliveryCost(@RequestBody OrderDto order);
+    BigDecimal deliveryCost(@RequestBody OrderDto order);
 }
